pyproject.toml
[tool.poetry]
name = "confluenpy"
version = "0.1.8"
description = ""
authors = ["Matteo Giani <matteo.giani.87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
atlassian-python-api = "^3.41.9"
pytest-mock = "^3.12.0"
validators = "^0.22.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
pytest = "^8.0.0"
isort = "^5.13.2"
black = "^24.1.1"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
python-dotenv = "^1.0.1"

[tool.black]
line-length = 150

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length = 150
disable = """
    redefined-outer-name
"""

[tool.pytest.ini_options]
addopts = """
-ra --strict-markers
"""

xfail_strict = true
testpaths = "tests"
markers = """
    unit: unit test
    integration: integration test
    end_to_end: end to end test
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
